src/requirements.txt
azure-common==1.1.11
azure-mgmt-containerinstance==0.4.0
azure-mgmt-resource==1.2.2